refactor: share product types
